requirements.txt
discord.py==0.11.0
setuptools==20.10.1
